requirements.txt
aws-cdk.core==1.133.0
aws-cdk.aws_iam==1.133.0
aws-cdk.aws_sqs==1.133.0
aws-cdk.aws_sns==1.133.0
aws-cdk.aws_sns_subscriptions==1.133.0
aws-cdk.aws_s3==1.133.0
aws-cdk.assertions==1.133.0
aws-cdk.aws-codepipeline==1.133.0
aws-cdk.aws-codepipeline-actions==1.133.0
aws-cdk.pipelines==1.133.0
aws-cdk.aws-dynamodb==1.133.0